+++ cargo-make/scripts/sqlx-prepare.sh
@@ -11,7 +11,7 @@ CRATES=(
   "tasker-orchestration"
   "tasker-worker"
   "tasker-client"
-  "tasker-cli"
+  "tasker-ctl"
   "tasker-pgmq"
   "workers/ruby/ext/tasker_core"
   "workers/rust"